Update of defaults
Documentation in README
Defaults updated to have app_user and app_group instead of specifying
app.user and app.group.
sudo_user set to app_user

Checking in module to make sure we are up to date.
No need of checking maven_repository and version if we are installing for the
first time.